[Small fixes include move type to a different location, remove wait from cypress, and link correct translation]
